Experimental session security
I'm attempting to have Laravel send the "Secure" flag with cookies while in the non-development environment (because we're not developing locally with HTTPS, and sending the flag locally with break sessions).

I have to push this to master because I have to test it on the server and I'm too lazy to alter the deployment script to clone from a branch.