Fix issue with preceeding zero bits being truncated
The removal of preceeding bits (which is either a JavaScript feature, or because of parseInt, or both) would cause every AND operation (used to calculate the number of matches) to always find at least one match. (Because every sequence would start with 1 after preceeding 0s were removed).

Fixed by manually adding 1 to the start of each sequence, and then deducting one from the lengths afterwards.